Much more explicit behaviour around conversions, callables, and some bug fixes.
Discovery now works quite a lot differently, whole-package auto-discovery is only used if the package is explicitly listed in 'discovery.packages', otherwise only explicitly listed callables are ever used, including methods. This avoids a whole load of potential weirdness in selecting the wrong callable because it might look right. Type conversions are also a lot less permissive now, again, so that this behaviour is more well defined and explicit, allowing the same level of freedom Morph previously gave, but by default being safe and predictable. Some bug fixes around things like bidirectional config resolution, etc.

+++ lab/planner/to/to.go
@@ -36,5 +36,5 @@ type Optional[T any] struct {
 
 type Nested struct {
 	Foo string
-	Bar int64
+	Bar int
 }
